refactor: create a check balance and put the check amount in the transfer folder
